lots of timing data, correctness testing, swapped nlohmann for rapidjson

+++ connector/cpp/client-sync/helpers.hpp
@@ -76,7 +76,7 @@ int str_to_intfield(std::string s, int ndigits) {
   // assert ndigits >= 0
   auto x = s.find(".");
   if (x != std::string::npos) {
-    return stoi(s) * (pow(10, ndigits));
+    return stof(s) * (pow(10, ndigits));
   } else {
     int span = s.size() - 1;
     while (s[span] == 0) {
@@ -85,6 +85,6 @@ int str_to_intfield(std::string s, int ndigits) {
     s.erase(span, s.size());
     auto n = s.size();
     // assert n - x - 1 <= ndigits, f'{n - x},{ndigits}'
-    return stoi(s.erase(x, 1)) * pow(10, (ndigits - n + x + 1));
+    return stof(s.erase(x, 1)) * pow(10, (ndigits - n + x + 1));
   }
 }
